requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
decouple==0.0.7
python-decouple==3.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.3.1
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.0
pytz==2025.2
Werkzeug==3.1.3
